[#1470] change definition of _ramstart from int to unsigned long
git-svn-id: svn://localhost/svn/uclinux-dist/trunk@3616 3e2a5524-39e8-4cb9-84a9-c49da69cba3a

+++ linux-2.6.x/include/asm-blackfin/bfin-global.h
@@ -103,7 +103,7 @@ extern unsigned long table_start, table_end;
 
 extern struct file_operations dpmc_fops;
 extern char _start;
-extern int _ramstart, _ramend, _rambase;
+extern unsigned long _ramstart, _ramend, _rambase;
 extern unsigned long memory_start, memory_end;
 extern unsigned long memory_mtd_end;
 extern char _stext_l1[], _etext_l1[], _sdata_l1[], _edata_l1[], _sbss_l1[],
